Extract FomRoleAdapterTrait for reuse
